♻️ refactor: Use useCarousel hook in Grid and List
- Moved common logic previously used in MediaSection to the useCarousel hook
